Use balto-utils NCC workflow to build dist
Building the distribution file that's actually used when this workflow
is referenced has been manual until now (and the current release's dist/
is out of sync with src/). But balto-utils has a workflow we can use to
automate that. Using the default setup for that workflow and hopefully
it works!